\name{getFfmpeg}
\alias{getFfmpeg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Download a Copy of Ffmpeg
}
\description{
Checks if ffmpeg is available, and if not, downloads a copy of Ffmpeg from https://www.gyan.dev and places the executable in the working directory. This executable can then be used for videomanipulation. 
}
\usage{
getFfmpeg()
}
\value{
Null
}
\author{
Jan Jaap Poos
}
